[Feature] Various changes, please read commit message
Fixed a discrepancy with cq->queue being a pointer, it should be
a value type not a pointer type.

Wrote a build script to compile and run the program (it's portable code
so it should run on your computer as well).

changed the infiniband directory to be lowercased.

added tests.h to avoid duplicate symbol errors

Signed-off-by: Amit Matityahu <amit2129@gmail.com>