[Feature] Changed variable sizes, enhanced Makefiles
Added status to cqe to signal failure to send.

Changed the variables sizes related to length and wr_id,
(increase from uint8_t to uint16_t).

Also fixed the process_send function to actually be able
to process a work request and send the data.

Central dispatch is still required, either as a part of
the infiniband subsystem or as an application layer.

Signed-off-by: Amit Matityahu <amit2129@gmail.com>